+++ quantify_scheduler/schedules/_visualization/constants.py
@@ -1,0 +1,20 @@
+# Repository: https://gitlab.com/quantify-os/quantify-scheduler
+# Licensed according to the LICENCE file on the main branch
+"""Constants used for drawing figures."""
+
+
+# draw circuit diagram mpl color scheme
+COLOR_BLUE = "#5296dd"
+COLOR_LAZURE = "#92bddf"
+COLOR_GREY = "#d7d7d7"
+COLOR_ORANGE = "#ff6314"
+
+# draw circuit diagram mpl scaling
+METER_WEDGE_ANGLE = 50
+METER_WEDGE_RADIUS = 0.4
+METER_WEDGE_HEIGHT_SCALING = 1.40
+
+METER_ARROW_ANGLE = 70
+METER_ARROW_HEIGHT_SCALING = 3.0
+
+ACQ_METER_LINEWIDTH = 0.5